requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.5
oauthlib==2.0.7
passlib==1.7.1
python-dateutil==2.6.1
requests==2.18.3
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.22
